reformat to allow app1 to view page
