+++ src/main/steps/respondent/review-the-application/template.njk
@@ -1,56 +1,1 @@
-{% extends "common/page.njk" %}
-
-{% block page_content %}
-  <h1 class="govuk-heading-l">{{ title }}</h1>
-  <p class="govuk-body">{{ line1 }}</p>
-  <h3 class="govuk-heading-m">{{ subHeading1 }}</h3>
-  <p class="govuk-body">{{ line2 | safe }}</p>
-  <p class="govuk-body">{{ line3 | safe }}</p>
-  <p class="govuk-body">{{ line4 | safe }}</p>
-  <p class="govuk-body">{{ line5 | safe }}</p>
-  <p class="govuk-body">{{ line6 }}</p>
-  <p class="govuk-body">{{ line7 | safe }}</p>
-  <p class="govuk-body">{{ line8 }}</p>
-  <p class="govuk-body">{{ line9 }}</p>
-  <h3 class="govuk-heading-m">{{ subHeading2 }}</h3>
-  <p class="govuk-body">{{ line10 }}</p>
-  <p class="govuk-body">{{ line11 | safe }}</p>
-  <p class="govuk-body">{{ line12 }}</p>
-  <p class="govuk-body">{{ line13 | safe }}</p>
-  <p class="govuk-body">{{ line14 }}</p>
-  <p class="govuk-body">{{ line15 | safe }}</p>
-  <p class="govuk-body">{{ line16 }}</p>
-  <h3 class="govuk-heading-m">{{ subHeading3 }}</h3>
-  <p class="govuk-body">{{ line17 }}</p>
-  <ul class="govuk-list govuk-list--bullet">
-    {% for line in connectionBulletPoints %}
-      <li>{{ line | safe }}</li>
-    {% endfor %}
-  </ul>
-  <p class="govuk-body">{{ jurisdictionsMoreDetails | safe }}</p>
-  <h3 class="govuk-heading-m">{{ subHeading4 }}</h3>
-  <p class="govuk-body">{{ line18 }}</p>
-  <p class="govuk-body">{{ line19 | safe }}</p>
-  <h3 class="govuk-heading-m">{{ subHeading5 }}</h3>
-  <p class="govuk-body">{{ line20 }}</p>
-  <h3 class="govuk-heading-m">{{ subHeading6 }}</h3>
-  <p class="govuk-body">{{ line21 }}</p>
-  <p class="govuk-body">{{ financialOrderMoreDetails | safe }}</p>
-  {% if isApplicantAddressNotPrivate %}
-    <h3 class="govuk-heading-m">{{ subHeading7 }}</h3>
-    <p class="govuk-body">{{ applicantAddressCountry | safe }}</p>
-    <h3 class="govuk-heading-m">{{ subHeading8 }}</h3>
-    <p class="govuk-body">{{ line22 | safe }}</p>
-  {% endif %}
-  {% if isRespondentAddressNotPrivate %}
-    <h3 class="govuk-heading-m">{{ subHeading9 }}</h3>
-    <p class="govuk-body">{{ respondentAddressCountry | safe }}</p>
-    <h3 class="govuk-heading-m">{{ subHeading10 }}</h3>
-    <p class="govuk-body">{{ line23 | safe }}</p>
-  {% endif %}
-  <h3 class="govuk-heading-m">{{ subHeading11 }}</h3>
-  <p class="govuk-body">{{ line24 | safe }}</p>
-  <p class="govuk-body">{{ applicantName | safe }}</p>
-  <h3 class="govuk-heading-m">{{ subHeading12 }}</h3>
-
-{% endblock %}
+{% include "../../applicant1/review-the-application/template.njk" %}
